frontend/requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
pydantic
openai
groq
tiktoken
jinja2
python-multipart
requests
